4 different implementation of as_html
calculating self.hands is a VERY SLOW operation. it needs to be cached!
so all as_html4() is doing is just caching the result of it, and instead
of 2 SECONDS, it only takes 24 milliseconds to calculate the HTML.
It's a pretty simple function, no need for a template engine like jinja2
because there is more logic in the template than HTML :)
Simple str concatenating will do it.